+++ code_coverage_html.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+./vendor/bin/phpunit --color --bootstrap vendor/autoload.php --coverage-html coverage --whitelist src tests