+++ test.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-./vendor/bin/phpunit --color --bootstrap vendor/autoload.php tests
+./vendor/bin/phpunit --color --bootstrap vendor/autoload.php --coverage-text --whitelist src tests