+++ Changelog.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] - 2020-04-21
+### Changed
+- Throw ArgumentException when not passing in an array or iterable.
+- Add `implode`
+- Add `filterValues`
+- Add `keys`
+- Add `values`
+### Added
+- 100% code coverage in tests
+
+## [0.3.0] - 2019-08-27
+### Changed
+- Allow iterable everywhere.
+- Make ArFluent immutable.
+### Added
+- Improve tests.
+
 ## [0.3.0] - 2019-08-27
 ### Changed
 - Allow iterable everywhere.